Add user2user and item2item predictions
Since the feature-vectors for each user and each anime essentially model
their features and their tastes in features, comparing animes and users
on that basis works quite well.
So Reco-Chan now generates a KD-Tree for users and animes, and when
looking for a similar user, simply searches for the k nearest neighbors.